refactor: flex and spacexy
